+++ src/main/java/songstress/cards/AbstractSongstressCard.java
@@ -1,15 +1,19 @@
 package songstress.cards;
 
 import com.megacrit.cardcrawl.core.CardCrawlGame;
+import com.megacrit.cardcrawl.dungeons.AbstractDungeon;
 import com.megacrit.cardcrawl.localization.CardStrings;
 
 import basemod.abstracts.CustomCard;
 import songstress.TheSongstressMod;
 import songstress.patches.ColorEnum;
+import songstress.powers.Holy;
+import songstress.powers.Muted;
 
 public abstract class AbstractSongstressCard extends CustomCard {
 
 	public boolean isHoly = false;
+	public int holyCost = 0;
 	public boolean isCure = false;
 	public boolean isSong = false;
 
@@ -29,6 +33,27 @@ public abstract class AbstractSongstressCard extends CustomCard {
 
 	public String getID() {
 		return cardID;
+	}
+
+	@Override
+	public boolean hasEnoughEnergy() {
+		if (Holy.holyAmount() < holyCost) {
+			cantUseMessage = Holy.NOT_ENOUGH_HOLY;
+			return false;
+		}
+		boolean isMuted = AbstractDungeon.player.hasPower(TheSongstressMod.withModID(Muted.POWER_ID));
+		if (isSong && isMuted) {
+			cantUseMessage = Muted.MUTED;
+			return false;
+		}
+		if (this instanceof DeafenedStab && !isMuted) {
+			cantUseMessage = Muted.NOT_MUTED;
+			return false;
+		}
+		if (!super.hasEnoughEnergy()) {
+			return false;
+		}
+		return true;
 	}
 
 }