+++ src/main/java/songstress/cards/SourNote.java
@@ -1,0 +1,60 @@
+package songstress.cards;
+
+import com.megacrit.cardcrawl.actions.AbstractGameAction;
+import com.megacrit.cardcrawl.actions.common.ApplyPowerAction;
+import com.megacrit.cardcrawl.actions.common.DamageAllEnemiesAction;
+import com.megacrit.cardcrawl.cards.AbstractCard;
+import com.megacrit.cardcrawl.characters.AbstractPlayer;
+import com.megacrit.cardcrawl.dungeons.AbstractDungeon;
+import com.megacrit.cardcrawl.localization.CardStrings;
+import com.megacrit.cardcrawl.monsters.AbstractMonster;
+import com.megacrit.cardcrawl.powers.WeakPower;
+
+public class SourNote extends AbstractSongstressCard {
+
+	public static final String ID = "SourNote";
+	private static final CardStrings cardStrings = getCardStrings(ID);
+	public static final String NAME = cardStrings.NAME;
+	private static final int COST = 1;
+	public static final String DESCRIPTION = cardStrings.DESCRIPTION;
+	private static final int WEAK_AMT = 2;
+	private static final int UPGRADE_WEAK = 1;
+	private static final int DMG = 7;
+	private static final int UPGRADE_DMG = 3;
+	private static final CardType TYPE = CardType.ATTACK;
+	private static final CardRarity RARITY = CardRarity.COMMON;
+	private static final CardTarget TARGET = CardTarget.ALL_ENEMY;
+
+	public SourNote() {
+		super(ID, NAME, COST, DESCRIPTION, TYPE, RARITY, TARGET);
+		magicNumber = baseMagicNumber = WEAK_AMT;
+		baseDamage = DMG;
+		isMultiDamage = true;
+		isSong = true;
+	}
+
+	@Override
+	public void upgrade() {
+		if (!upgraded) {
+			upgradeName();
+			upgradeMagicNumber(UPGRADE_WEAK);
+			upgradeDamage(UPGRADE_DMG);
+		}
+	}
+
+	@Override
+	public AbstractCard makeCopy() {
+		return new SourNote();
+	}
+
+	@Override
+	public void use(AbstractPlayer player, AbstractMonster monster) {
+		AbstractDungeon.actionManager.addToBottom(new DamageAllEnemiesAction(player, multiDamage, damageTypeForTurn,
+				AbstractGameAction.AttackEffect.SLASH_HORIZONTAL));
+		for (AbstractMonster mo : AbstractDungeon.getCurrRoom().monsters.monsters) {
+			AbstractDungeon.actionManager.addToBottom(new ApplyPowerAction(mo, player,
+					new WeakPower(mo, magicNumber, false), magicNumber, true, AbstractGameAction.AttackEffect.NONE));
+		}
+	}
+
+}